chore: sync remaining billing portal updates
- Add Internal.tsx component for billing portal
- Remove error.tsx (now handled by Internal component)
- Update loading states and page structure
- Sync auth context and database client updates
- Update registry types for invoices and subscriptions
