+++ platform/flowglad-next/src/app/billing-portal/[organizationId]/[customerId]/loading.tsx
@@ -1,6 +1,6 @@
 export default function BillingPortalLoading() {
   return (
-    <div className="min-h-screen bg-background">
+    <div className="bg-background">
       {/* Header skeleton */}
       <header className="border-b bg-background/95 backdrop-blur supports-[backdrop-filter]:bg-background/60 sticky top-0 z-50">
         <div className="container mx-auto px-4 py-4 max-w-6xl">
