+++ platform/flowglad-next/src/app/billing-portal/layout.tsx
@@ -1,4 +1,7 @@
 import { Metadata } from 'next'
+import Providers from '../Providers'
+import { getSession } from '@/utils/auth'
+import { betterAuthUserToApplicationUser } from '@/utils/authHelpers'
 
 export const metadata: Metadata = {
   title: 'Billing Portal - Flowglad',
@@ -6,10 +9,25 @@ export const metadata: Metadata = {
     'Manage your subscription, payment methods, and invoices',
 }
 
-export default function BillingPortalLayout({
+export default async function BillingPortalLayout({
   children,
 }: {
   children: React.ReactNode
 }) {
-  return <>{children}</>
+  const session = await getSession()
+
+  return (
+    <Providers
+      authContext={{
+        organization: undefined,
+        livemode: false,
+        user: session
+          ? await betterAuthUserToApplicationUser(session?.user)
+          : undefined,
+        role: 'customer',
+      }}
+    >
+      {children}
+    </Providers>
+  )
 }
